Refine 'Value of BEAR.Sunday' section in technical documentation
This commit streamlines the section discussing the value of BEAR.Sunday in our technical documentation. It better articulates how this framework contributes to the productivity, maintainability, and versatility of the software for developers. The user benefits in terms of performance, reliability, and ease of use are more focused, and the added-value for businesses is succinctly highlighted.